Level 2 Boss Markers
Play-testing showed that level 2 is large in scope which can lead to users not reaching the boss within the allotted 10 minutes of playtime. To address this I've added some arrows on the road to point players in the right direction towards the boss.
this can be a temporary fix to replaced with a more elegant solution if time allows.